Rewritten with continuations.
The script has been rewritten into two parts. A "Retriever" which
fetches the most recent tweet from the queue, and a "Tweeter" which
takes that tweet and posts it to the main account. This new strategy
uses PyPy's continulets to suspend computation in the Tweeter, while the
Retriever is fetching a tweet, and then switching back to the Tweeter.

One new behavior is that the Tweeter will waith 60 minutes after the
most recent tweet, and when that time is up, it will switch to the
Retriever which will continuously poll twitter for new tweets. Once it
gets one, it will switch back, and the Tweeter will immediately post it.