Fixes bug causing tweets to be posted at night.
Before the timeToTweet cutoff, butterbird would sometimes enter the
tweet retrieval loop and spin waiting for a tweet to show up in the
timeline. This would go on past the timeToTweet cutoff, but when a tweet
finally did show up, it would be retrieved and posted even though it was
no longer time to tweet. Now an additional check is inserted after the
tweet is retrieved to prevent this from happening.